Add support for numeric delete-after values
Delte-after can be a string (eg. "never") or an integer. RabbitMQ does
not accept string-formatted integers ("42") so this PR changes the type
of the `DeleteAfter` field and adds a custom marshaler/unmarshaler
to handle both cases. As you can see in the tests, this may be a
breaking change - since values of different types cannot be equal in Go,
when comparing `DeleteAfter` with a string, you now need to convert one of the
values. However, most operations should work without code changes.

Fixes #149