Make tests more resilient to slow emit stats
Without the `eval` instructions, the stats were emitted "too slowly",
and some tests expect metrics to be available sooner. The update to the
Makefile ensures that stats are emitted "quickly". Those instructions
are borrowed from the CI scripts in bin/

The expectations in the test prevent a panic when the stats are not
emitted fast enough.